Split tests in two files
